requirements.txt
streamlit>=1.40.0
google-generativeai>=0.8.0
Pillow>=10.0.0
tenacity>=8.0.0
python-dotenv>=1.0.0
streamlit-paste-button>=0.1.0
requests>=2.31.0
pandas>=2.0.0
